+++ packages/ui/src/components/molecules/InterviewAnswerResultCard/interviewAnswerResultCard.css.ts
