+++ packages/ui/src/components/molecules/InterviewScoreCard/interviewScoreCard.css.ts
