+++ packages/ui/src/components/molecules/SelectSubjectCard/selectSubjectCard.css.ts
