+++ packages/ui/src/components/molecules/SelectSubjectCard/selectSubjectCard.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import SelectSubjectCard from '@/components/molecules/SelectSubjectCard/SelectSubjectCard';
+import SelectSubjectCard from '@/components/molecules/SelectSubjectCard';
 import { SUBJECT_OPTIONS } from '@/components/molecules/SelectSubjectCard/subjectOptions';
 
 const meta = {
